薪酬发放优化逻辑

+++ DatabaseProject/HR_java/src/main/java/com/hr_java/serviceImpl/SerialServiceImpl.java
@@ -61,11 +61,19 @@ public class SerialServiceImpl extends ServiceImpl<SerialMapper, Serial> impleme
             System.err.println("输入的数值不允许为负数！");
             return false;
         }
-        return getBaseMapper().updateSerialByID(serialID,bounty,penalty);
+        if(!"正常".equals(getBaseMapper().getStatus(serialID))){ //待审核，驳回状态才能修改内部条目
+            return getBaseMapper().updateSerialByID(serialID,bounty,penalty);
+        }else{
+            return false;
+        }
     }
 
     @Override
     public boolean deleteSerialByID(Integer serialID) {
-        return getBaseMapper().deleteSerialByID(serialID);
+        if(!"正常".equals(getBaseMapper().getStatus(serialID))){//待审核，驳回状态才能修改内部条目
+            return getBaseMapper().deleteSerialByID(serialID);
+        }else{
+            return false;
+        }
     }
 }
